Add ZTS and PHP 7.3 GH action builds for Windows
That makes a matrix of 12 jobs!  While PHP 8.0 support is desirable and
possible, there is a bit of an issue regarding the build environment
(windows-2016 vs. windows-2019).  Instead of diving deeper into this,
we should have a look at building the PHP 7 versions of the extension
with Visual Studio 2019, what is supposed to be possible, but not (yet)
directly supported by the PHP SDK.  This would also prevent failing
builds when windows-2016 will finally be retired.

Furthermore, likely more pressing needs would be caching and supporting
local usage of the Powershell scripts.