Fixed bug #81552 (xdiff_string_bpatch() and others missing)
Added stubs for the actual names of the functions for which aliases
exist. These are the same except that the names are changed and @alias
is removed from the doc comments.

Regenerated xdiff_arginfo.h. I had to remove the unnecessary PHPDoc
types in order for this to succeed using the PHP 8.1 version of
gen_stub.php. (PHP 8.0's version of the script doesn't complain.)